HDV-4 Restructured files into a CMake system
