+++ .gitignore
@@ -1,0 +1,3 @@
+Build/
+bin/
+Testing/
