+++ Source/CMakelists.txt
@@ -1,0 +1,17 @@
+cmake_minimum_required(VERSION 3.4)
+
+project(HDV VERSION 1.0)
+
+include(CTest)
+
+include_directories(include)
+add_subdirectory(include)
+add_subdirectory(src)
+add_subdirectory(app)
+
+# Conditionally add tests
+if (BUILD_TESTING)
+  add_subdirectory(tests)
+endif()
+
+install(TARGETS HDV DESTINATION bin)
