+++ Source/include/CMakelists.txt
@@ -1,0 +1,1 @@
+configure_file(hdv_config.h.in hdv_config.h)
