+++ Source/tests/test_cmakeBuildDef.cpp
@@ -1,0 +1,4 @@
+int main()
+{
+  return 0;
+}
